+++ agent/lm_agent/config.py
@@ -60,8 +60,11 @@ class Settings(BaseSettings):
     # path to the license server features config file
     LICENSE_SERVER_FEATURES_CONFIG_PATH: Optional[str]
 
-    # a path to a folder containing binaries for license management tools
-    BIN_PATH: DirectoryPath = _DEFAULT_BIN_PATH
+    # path to the binary for lmstat (needed for FlexLM licenses)
+    LMSTAT_BIN_PATH: DirectoryPath = _DEFAULT_BIN_PATH
+
+    # path to the binary for rlmstat (needed for RLM licenses)
+    RLMSTAT_BIN_PATH: DirectoryPath = _DEFAULT_BIN_PATH
 
     # debug mode turns on certain dangerous operations
     DEBUG: bool = False